Fix budget size

+++ decidim-budgets/lib/decidim/api/mutations/budget_attributes.rb
@@ -9,7 +9,7 @@ module Decidim
       argument :description, GraphQL::Types::JSON, description: "Description of the budget", required: false
       argument :title, GraphQL::Types::JSON, description: "title of the budget", required: false
       argument :total_budget, GraphQL::Types::Int, description: "Total budget of the budget", required: false
-      argument :weight, GraphQL::Types::Int, description: "weight of the budget", required: false
+      argument :weight, GraphQL::Types::BigInt, description: "weight of the budget", required: false
     end
   end
 end
